+++ pkg/constraints/constraint_generator.go
@@ -3,8 +3,8 @@ package constraints
 import (
 	"context"
 
-	"github.com/operator-framework/deppy/internal/entitysource"
-	"github.com/operator-framework/deppy/internal/sat"
+	"github.com/operator-framework/deppy/pkg/entitysource"
+	"github.com/operator-framework/deppy/pkg/sat"
 )
 
 // ConstraintGenerator generates solver constraints given an entity querier interface
