+++ pkg/sat/bench_test.go
