+++ pkg/sat/doc.go
