+++ pkg/sat/solve.go
