+++ pkg/sat/tracer.go
