+++ pkg/solver/solver_suite_test.go
